requirements.txt
# Garden AI - image_generation (BFL FLUX Fill PRO)
# Python 3.9+

# API BFL
requests>=2.28.0

# Images
Pillow>=9.0.0
numpy>=1.21.0

# UI (optionnel)
streamlit>=1.28.0

# Optionnel (améliore le masque plantable)
scipy>=1.7.0

# Prérequis : export BFL_API_KEY='votre_clé' (ou MOCK_BFL=true)
